scraper/requirements.txt
python-jobspy
pandas
tqdm
pyyaml
beautifulsoup4
